+++ code/GitInsight/GitInsight.csproj
