+++ code/GitInsight/bin/Debug/net6.0/GitInsight.deps.json
@@ -80,7 +80,7 @@
           "runtimes/win-x64/native/git2-106a5f2.dll": {
             "rid": "win-x64",
             "assetType": "native",
-            "fileVersion": "0.0.0.0"
+            "fileVersion": "0.28.4.0"
           },
           "runtimes/win-x64/native/git2-106a5f2.pdb": {
             "rid": "win-x64",
@@ -90,7 +90,7 @@
           "runtimes/win-x86/native/git2-106a5f2.dll": {
             "rid": "win-x86",
             "assetType": "native",
-            "fileVersion": "0.0.0.0"
+            "fileVersion": "0.28.4.0"
           },
           "runtimes/win-x86/native/git2-106a5f2.pdb": {
             "rid": "win-x86",
